+++ examples/when.rb
@@ -29,4 +29,19 @@ raise "got_start not set" unless got_start
 raise "got_start_when_given_as_symbol not set" unless got_start_when_given_as_symbol
 raise "got_exit not set" unless got_exit
 
+k = Kommando.new "not_existing_command_with non_existing_args"
+
+got_error = false
+k.when :error do
+  puts ":error"
+  got_error = true
+end
+
+begin
+  k.run
+rescue
+end
+
+raise "got_error not set" unless got_error
+
 puts "end"